requirements.txt
streamlit
PyPDF2
google.generativeai
python-dotenv
pdf2image
fpdf 
language-tool-python
reportlab
pytesseract
matplotlib
poppler-utils
